Fix Publish All for sites without published repo (#3006)
* Fix Publish All for sites without published repo

* Updates according to feedback

+++ src/main/java/org/craftercms/studio/impl/v2/repository/GitContentRepository.java
@@ -49,6 +49,7 @@ import org.craftercms.studio.api.v1.to.DeploymentItemTO;
 import org.craftercms.studio.api.v1.util.filter.DmFilterWrapper;
 import org.craftercms.studio.api.v2.annotation.RetryingOperation;
 import org.craftercms.studio.api.v2.dal.*;
+import org.craftercms.studio.api.v2.exception.PublishedRepositoryNotFoundException;
 import org.craftercms.studio.api.v2.repository.ContentRepository;
 import org.craftercms.studio.api.v2.repository.RepositoryChanges;
 import org.craftercms.studio.api.v2.repository.RetryingRepositoryOperationFacade;
@@ -1147,6 +1148,10 @@ public class GitContentRepository implements ContentRepository, DeploymentHistor
         retryingRepositoryOperationFacade.call(deleteCommand);
     }
 
+    protected boolean branchExists(Repository repo, String branch) throws IOException {
+        return repo.resolve(branch) != null;
+    }
+
     private void cleanUpMoveFolders(Git git, String path) throws GitAPIException, IOException {
         Path parentToDelete = Paths.get(path).getParent();
         boolean isPage = path.endsWith(FILE_SEPARATOR + INDEX_FILE);
@@ -1795,15 +1800,9 @@ public class GitContentRepository implements ContentRepository, DeploymentHistor
                                         userServiceInternal, encryptor, generalLockService,
                                         retryingRepositoryOperationFacade);
         Repository repo = helper.getRepository(siteId, GitRepositories.PUBLISHED);
-        // if the published repo doesn't exist yet, create it and finish since it will be in sync already
+        // if the published repo doesn't exist yet, fail
         if (repo == null) {
-            logger.info("Creating published repository for site {0}", siteId);
-            if (helper.createPublishedRepository(siteId, site.getSandboxBranch())) {
-                // return an empty set so no additional operations are performed
-                return new RepositoryChanges();
-            } else {
-                throw new ServiceLayerException("Error creating published repository for site " + siteId);
-            }
+            throw new PublishedRepositoryNotFoundException("Published repository not found for site " + siteId);
         }
         String repoLockKey = helper.getPublishedRepoLockKey(siteId);
         generalLockService.lock(repoLockKey);
@@ -1815,6 +1814,12 @@ public class GitContentRepository implements ContentRepository, DeploymentHistor
             retryingRepositoryOperationFacade.call(git.pull()
                     .setRemote(DEFAULT_REMOTE_NAME)
                     .setStrategy(THEIRS));
+            // check if the target branch exists
+            if (!branchExists(repo, publishingTarget)) {
+                throw new PublishedRepositoryNotFoundException("Publishing target branch " + publishingTarget +
+                                                               " not found for site " + siteId);
+            }
+
             // checkout target branch
             checkoutBranch(git, publishingTarget);
 
@@ -1866,8 +1871,6 @@ public class GitContentRepository implements ContentRepository, DeploymentHistor
 
             return new RepositoryChanges(updatedPaths, deletedPaths);
         } catch (GitAPIException | IOException e) {
-            // unlock the repo in case of error to prevent deadlocks
-            generalLockService.unlock(repoLockKey);
             throw new ServiceLayerException("Error publishing all changes for site " + siteId + " to target " +
                     publishingTarget, e);
         }
@@ -1924,13 +1927,19 @@ public class GitContentRepository implements ContentRepository, DeploymentHistor
                                         retryingRepositoryOperationFacade);
         String repoLockKey = helper.getPublishedRepoLockKey(siteId);
         Repository repo = helper.getRepository(siteId, GitRepositories.PUBLISHED);
+        if (repo == null) {
+            return;
+        }
         try (Git git = Git.wrap(repo)) {
             resetIfNeeded(repo, git);
 
-            checkoutBranch(git, publishingTarget);
+            // go batch to the target branch if it exists
+            if (branchExists(repo, publishingTarget)) {
+                checkoutBranch(git, publishingTarget);
 
-            String inProgressBranchName = publishingTarget + IN_PROGRESS_BRANCH_NAME_SUFFIX;
-            deleteBranches(git, inProgressBranchName);
+                String inProgressBranchName = publishingTarget + IN_PROGRESS_BRANCH_NAME_SUFFIX;
+                deleteBranches(git, inProgressBranchName);
+            }
         } catch (GitAPIException | IOException e) {
             throw new ServiceLayerException("Error canceling all changes for site " + siteId + " to target " +
                     publishingTarget, e);
